+++ src/features/babel/index.ts
@@ -1,11 +1,17 @@
 import type { FeatureSetup, IsSkipFeature } from '../../types'
 import { resolve } from 'path'
 import { BUILD_TOOLS } from '../typescript/build-tools'
-import { addDevDeps } from '../../core/dependency'
+import { addDep, addDevDeps } from '../../core/dependency'
 import { rederTemplate } from '../../core/template'
+import { fileExists } from '../../utils/path_helper'
 
-export const isSkip: IsSkipFeature = async ({ answers }) => {
-  return [BUILD_TOOLS.TSC, BUILD_TOOLS.ESBUILD].includes(answers.buildTool)
+const configFileExists = async (path: string) => await fileExists(resolve(path, '.babelrc'))
+
+export const isSkip: IsSkipFeature = async ({ rootPath, answers }) => {
+  return (
+    (await configFileExists(rootPath)) ||
+    [BUILD_TOOLS.TSC, BUILD_TOOLS.ESBUILD].includes(answers.buildTool)
+  )
 }
 
 export const setup: FeatureSetup = async context => {
@@ -18,8 +24,11 @@ export const setup: FeatureSetup = async context => {
     '@babel/preset-env',
     '@babel/plugin-proposal-class-properties',
     '@babel/plugin-proposal-object-rest-spread',
-    '@babel/preset-typescript'
+    '@babel/preset-typescript',
+    '@babel/plugin-transform-runtime'
   ])
+
+  addDep('@babel/runtime')
 
   if (isReactNeeded) {
     addDevDeps(['@babel/preset-react'])