+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# V0.2.1
+
+- Renew rollup config
+- Added runtim plugin for babel
+
 # V0.2.0
 
 - Added esbuild feature for build typescript